phys_to_mach and mach_to_phys tables contain long entries, not 32-bit entries.

+++ linux-2.6-xen-sparse/arch/xen/x86_64/kernel/setup.c
@@ -76,7 +76,7 @@ EXPORT_SYMBOL(HYPERVISOR_shared_info);
 /* Allows setting of maximum possible memory size  */
 unsigned long xen_override_max_pfn;
 
-u32 *phys_to_machine_mapping, *pfn_to_mfn_frame_list;
+unsigned long *phys_to_machine_mapping, *pfn_to_mfn_frame_list;
 
 EXPORT_SYMBOL(phys_to_machine_mapping);
 
@@ -734,22 +734,22 @@ void __init setup_arch(char **cmdline_p)
 		/* Make sure we have a large enough P->M table. */
 		if (end_pfn > xen_start_info->nr_pages) {
 			phys_to_machine_mapping = alloc_bootmem(
-				end_pfn * sizeof(u32));
+				end_pfn * sizeof(unsigned long));
 			memset(phys_to_machine_mapping, ~0,
-			       end_pfn * sizeof(u32));
+			       end_pfn * sizeof(unsigned long));
 			memcpy(phys_to_machine_mapping,
-			       (u32 *)xen_start_info->mfn_list,
-			       xen_start_info->nr_pages * sizeof(u32));
+			       (unsigned long *)xen_start_info->mfn_list,
+			       xen_start_info->nr_pages * sizeof(unsigned long));
 			free_bootmem(
 				__pa(xen_start_info->mfn_list), 
 				PFN_PHYS(PFN_UP(xen_start_info->nr_pages *
-						sizeof(u32))));
+						sizeof(unsigned long))));
 		}
 
 		pfn_to_mfn_frame_list = alloc_bootmem(PAGE_SIZE);
 
-		for ( i=0, j=0; i < end_pfn; i+=(PAGE_SIZE/sizeof(u32)), j++ )
-			pfn_to_mfn_frame_list[j] = 
+		for ( i=0, j=0; i < end_pfn; i+=(PAGE_SIZE/sizeof(unsigned long)), j++ )
+			pfn_to_mfn_frame_list[j] =
 				virt_to_mfn(&phys_to_machine_mapping[i]);
 	}
 #endif
